* extract superclass for RN promise
